Refactor Postgres gateways to remove unnecessary interface and drop AWS
