--- src/shared-types/src/BichardPostgresGateway.ts
@@ -1,6 +1,0 @@
-import type { PromiseResult } from "./Result"
-
-export default interface BichardPostgresGateway {
-  deleteArchivedErrors(messageId: string): PromiseResult<void>
-  messageIsInErrorList(messageId: string): PromiseResult<boolean>
-}
